requirements.txt
anyio==3.5.0
asgiref==3.5.2
astroid==2.11.2
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
dill==0.3.4
Django==4.1.3
djangorestframework==3.14.0
dnspython==2.2.1
docopt==0.6.2
email-validator==1.3.0
entrypoints==0.4
executing==0.8.3
fastapi==0.85.0
h11==0.12.0
httptools==0.2.0
idna==3.3
ipykernel==6.13.0
ipython==8.2.0
isort==5.10.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.2.2
jupyter-core==4.9.2
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.7.0
nest-asyncio==1.5.5
orjson==3.8.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==1.0.18
psutil==5.9.0
pure-eval==0.2.2
pycodestyle==2.8.0
pydantic==1.9.0
Pygments==2.11.2
pylint==2.13.5
pyparsing==3.0.8
pyte==0.5.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2022.6
pywin32==303
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.3
stack-data==0.2.0
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2022.6
ujson==5.5.0
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.1
wrapt==1.14.0
